requirements.txt
django==1.11
django-allauth==0.33.0
psycopg2==2.7.3.1
djangorestframework==3.6.4
Pillow==4.2.1
